Added new datastructure and NB-IoT functionality
