--- board_support_crate/aau_ubiss/__init__.py
@@ -1,1 +1,0 @@
-from ._core import aau_ubiss
