+++ board_support_crate/aauiot/__init__.py
@@ -1,0 +1,1 @@
+from ._core import aau_iot, MqttData
