+++ board_support_crate/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "AAU_Ubiss"
-version = "0.3.2"
+name = "aau-iot-testbed"
+version = "0.4.2"
 authors = [
   { name="Jonas Jakobsen", email="jojako21@student.aau.dk" },
   { name="Andreas Casparsen", email="aca@es.aau.dk"},
@@ -19,6 +19,7 @@ dependencies = [
   "adafruit-circuitpython-sgp30",
   "lgpio",
   "paho-mqtt",
-  "pytz"
+  "pytz",
+  "pyserial"
 ]
